# docs/swagger.yaml
basePath: /
definitions:
  main.Item:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a simple API for managing items.
  title: Item API
  version: "1.0"
paths:
  /items:
    get:
      description: Retrieve all items in the list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Item'
            type: array
      summary: Get all items
      tags:
      - items
    post:
      consumes:
      - application/json
      description: Add a new item to the list
      parameters:
      - description: New Item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/main.Item'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Item'
      summary: Create an item
      tags:
      - items
  /items/{id}:
    delete:
      description: Remove an item by ID
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Item not found
          schema:
            type: string
      summary: Delete an item
      tags:
      - items
    get:
      description: Retrieve a single item by ID
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Item'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Item not found
          schema:
            type: string
      summary: Get an item by ID
      tags:
      - items
    put:
      consumes:
      - application/json
      description: Update an existing item by ID
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/main.Item'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Item'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Item not found
          schema:
            type: string
      summary: Update an item
      tags:
      - items
swagger: "2.0"